Separate header of buffer fusing
